Brave New Spooler - streams done right
We've now moved refilling logic out of the FileStreams
themselves, and into a separate thread that spools audio data
for them (using some nice crossbeam lockless data structures).

This is the proper way to do things, as IOing in the audio thread
is not something the PortAudio API recommends.

This has the added benefit of making things slightly simpler, as
the spooler handles quite a bit of shared state and just sends
samples to the FileStreams for callback-ing. Woot!