The Great De-Cruftification - remove RW contexts, implement messages
This commit gets rid of the last commit's crappy Action system, and
the terrible ReadableContext/WritableContext thing. Eugh. Now, the backend
acts more like a server, fielding Messages that the frontend sends it.
The frontend doesn't ever actually get to touch the commands themselves
(vastly simplifying things) - it gets CommandDescriptors that do the same job.

This implementation sucks and is only enough to execute a few commands to
show that it works. Don't use this!